requirement.txt
loguru==0.7.2
python-dotenv==1.0.1
python-telegram-bot==21.1
requests==2.31.0
urllib3==1.26.15
validators==0.28.0
